refactor(api): split room routes into handlers/routes/index pattern
Follow harness-docs presentation layer convention:

- room.handlers.ts: DO stub resolution + request handling logic
- room.routes.ts: Hono route registration (thin, calls handlers)
- room.index.ts: barrel re-export for the room router
- shared.ts: Env type, getDurableObjectStub, generateRoomCode

Each presentation module now has a clear separation:
routes = wiring, handlers = logic, index = public API.

Made-with: Cursor
