+++ apps/api/src/presentation/http/room/room.index.ts
@@ -1,0 +1,1 @@
+export { createRoomRouter } from "./room.routes";
